fix: Creates testing logic for the Rundown View and makes the rundown tests fit again
