+++ src/app/rundown-view/components/rundown/rundown.component.spec.ts
@@ -1,7 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { RundownComponent } from './rundown.component'
 import { anyString, anything, instance, mock, when } from '@typestrong/ts-mockito'
-import { ActivatedRoute, ActivatedRouteSnapshot, convertToParamMap, ParamMap, RouterModule } from '@angular/router'
 import { RundownStateService } from '../../../core/services/rundown-state.service'
 import { Subscription } from 'rxjs'
 import { RundownService } from '../../../core/abstractions/rundown.service'
@@ -22,9 +21,7 @@ async function configureTestBed(
   const mockedRundownStateService = params.mockedRundownStateService ?? createMockOfRundownStateService()
 
   await TestBed.configureTestingModule({
-    imports: [RouterModule.forRoot([])],
     providers: [
-      { provide: ActivatedRoute, useValue: instance(createMockOfActivatedRoute()) },
       { provide: RundownService, useValue: instance(mockedRundownService) },
       { provide: RundownStateService, useValue: instance(mockedRundownStateService) },
       { provide: DialogService, useValue: instance(mockedDialogService) },
@@ -43,15 +40,4 @@ function createMockOfRundownStateService(): RundownStateService {
   const mockedSubscription = mock<Subscription>()
   when(mockedRundownStateService.subscribeToRundown(anyString(), anything())).thenResolve(instance(mockedSubscription))
   return mockedRundownStateService
-}
-
-function createMockOfActivatedRoute(params: { paramMap?: ParamMap } = {}): ActivatedRoute {
-  const paramMap: ParamMap = params.paramMap ?? convertToParamMap({ rundownId: 'some-rundown-id' })
-  const mockedActivatedRouteSnapshot: ActivatedRouteSnapshot = mock<ActivatedRouteSnapshot>()
-  when(mockedActivatedRouteSnapshot.paramMap).thenReturn(paramMap)
-
-  const mockedActivatedRoute: ActivatedRoute = mock<ActivatedRoute>()
-  when(mockedActivatedRoute.snapshot).thenReturn(instance(mockedActivatedRouteSnapshot))
-
-  return mockedActivatedRoute
 }
