+++ src/app/rundown-view/rundown-view-routes.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core'
 import { RouterModule, Routes } from '@angular/router'
-import { RundownComponent } from './components/rundown/rundown.component'
 import { RundownViewComponent } from './components/rundown-view/rundown-view.component'
 
 const routes: Routes = [
