+++ src/app/rundown-view/rundown-view.module.ts
@@ -19,7 +19,7 @@ import { OffsetablePieceComponent } from './components/offsetable-piece/offsetab
 import { EditorialLineComponent } from './components/editorial-line/editorial-line.component'
 import { MatCardModule } from '@angular/material/card'
 import { ProducerShelfModule } from '../producer-shelf/producer-shelf.module'
-import { RundownHeaderPanelComponent } from './components/rundown-header-panel/rundown-header-panel.component';
+import { RundownHeaderPanelComponent } from './components/rundown-header-panel/rundown-header-panel.component'
 import { RundownViewComponent } from './components/rundown-view/rundown-view.component'
 
 @NgModule({
